requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
joblib==1.0.1
Markdown==3.3.4
nltk==3.6.2
packaging==20.9
protobuf==3.17.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
rake-nltk==1.0.4
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
uritemplate==3.0.1
urllib3==1.26.5
